Add AudioCompression (ulaw)

+++ app/src/main/java/de/rochefort/childmonitor/AudioListener.java
@@ -17,5 +17,5 @@
 package de.rochefort.childmonitor;
 
 public interface AudioListener {
-    void onAudio(byte[] audioBytes);
+    void onAudio(short[] audioBytes);
 }
